+++ packages/typescript/monorepo-pnpm-vite/boilerplate/packages/global.d.ts
@@ -1,0 +1,7 @@
+// global types
+
+declare namespace NodeJS {
+  interface ProcessEnv {
+    NODE_ENV: 'development' | 'production' | 'test'
+  }
+}