+++ packages/typescript/monorepo-pnpm-vite/boilerplate/packages/main/src/index.ts
@@ -1,0 +1,3 @@
+import { sum } from '@my/shared'
+
+console.log(sum(1, 2))