+++ packages/typescript/monorepo-pnpm-vite/boilerplate/packages/shared/src/__tests__/index.spec.ts
@@ -1,0 +1,8 @@
+import { expect, it } from 'vitest'
+import { sum } from '..'
+
+it('should return the sum of two numbers', () => {
+  expect(sum(2, 3)).toBe(5)
+  expect(sum(10, 5)).toBe(15)
+  expect(sum(-2, 5)).toBe(3)
+})